--- intro-me-domain/src/main/java/com/introme/user/entity/OAuthAttributes.java
@@ -1,42 +1,0 @@
-package com.introme.user.entity;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.Map;
-
-@Getter
-public class OAuthAttributes {
-    private Map<String, Object> attributes;
-    private String nameAttributeKey;
-    private String name;
-    private String email;
-
-    @Builder
-    public OAuthAttributes(Map<String, Object> attributes, String nameAttributeKey, String name, String email) {
-        this.attributes = attributes;
-        this.nameAttributeKey = nameAttributeKey;
-        this.name = name;
-        this.email = email;
-    }
-
-    public static OAuthAttributes of(String usernameAttributeName, Map<String, Object> attributes) {
-        return ofGoogle(usernameAttributeName, attributes);
-    }
-
-    private static OAuthAttributes ofGoogle(String usernameAttributeName, Map<String, Object> attributes) {
-        return OAuthAttributes.builder()
-                .name((String) attributes.get("name"))
-                .email((String) attributes.get("email"))
-                .attributes(attributes)
-                .nameAttributeKey(usernameAttributeName)
-                .build();
-    }
-    public IntroMeUser toEntity() {
-        return IntroMeUser.builder()
-                .name(name)
-                .email(email)
-                .role(Role.User)
-                .build();
-    }
-}
